Modificado Juego
-Creada clase pruebaRellenarEjercitosPrueba
- Generado metodo en juego el cual mediante constructor se pasan la lista de batallones de ambos ejercitos y estos se rellenan automaticamente
